+++ src/FolkerKinzel.Uris/md/ReleaseNotes.md
@@ -1,5 +1,22 @@
-﻿- Dependency update.
+﻿- Dependency update
+- The method `static StringBuilder AppendEmbeddedFileTo(this StringBuilder, string, string?, DataEncoding)` is not an extesion method anymore.
+- The `mimeTypeString` parameters in the `DataUrl` class have been renamed to `mimeType`.
 
+- New methods:
+```csharp
+StringBuilder DataUrl.AppendEmbeddedBytesTo(StringBuilder, byte[]?, in MimeTypeInfo, DataEncoding);
+StringBuilder DataUrl.AppendEmbeddedFileTo(StringBuilder, string, in MimeTypeInfo, DataEncoding dataEncoding);
+string DataUrl.FromFile(string, in MimeTypeInfo mimeType, DataEncoding);
+string DataUrl.FromBytes(byte[]?, in MimeTypeInfo, DataEncoding);
+```
+- Obsolete methods:
+```csharp
+StringBuilder DataUrl.AppendEmbeddedBytesTo(StringBuilder, byte[]?, MimeType, DataEncoding);
+StringBuilder DataUrl.AppendEmbeddedFileTo(this StringBuilder, string, MimeType, DataEncoding);
+string DataUrl.FromFile(string, MimeType mimeType, DataEncoding);
+string DataUrl.FromBytes(byte[]?, MimeType, DataEncoding);
+```
+.
 >**Notes to users who build .NET Core 2.x/3.0 applications based on the .NET Standard 2.0 part of the package:**  
 >You might get a compiler error. This is caused by a Microsoft dependency. You can get rid of this error, if you copy `<SuppressTfmSupportBuildWarnings>true</SuppressTfmSupportBuildWarnings>` to a `<PropertyGroup>` of your project file (at own risk).